Fixed: - Mesh behaviour on player. - Collider on phantoms. - Warning regarding playerMesh alpha transparency.
Added:
- Lamp damage, length and width increasing based on ScoreConfig.asset

+++ Assets/_/Tools/Shared/SO/LampConfig.cs
@@ -6,8 +6,8 @@ namespace Tools
     public class LampConfig : ScriptableObject
     {
 
-        public int m_damage;
-        public int m_lenght;
-        public int m_width;
+        public float m_damage;
+        public float m_lenght;
+        public float m_width;
     }
 }
